==> notebooks/12_inferencia_prod.ipynb <==
## RETO 1: Cargar Modelo en Producción

from pyspark.sql import SparkSession
from pyspark.sql.functions import col, current_timestamp, avg, count, when
import mlflow
import mlflow.spark
import os

spark = SparkSession.builder \
    .appName("SECOP_Produccion_Final") \
    .master("spark://spark-master:7077") \
    .getOrCreate()

print(" Sesión de Spark lista.")

from pyspark.ml.classification import LogisticRegression
from pyspark.ml.feature import Binarizer
from pyspark.ml import Pipeline

model_name = "Clasificador_Contratos_Top25"

try:
    print(" Intentando cargar desde MLflow Registry...")
    
    model_uri = f"models:/{model_name}/Production"
    production_model = mlflow.spark.load_model(model_uri)
    print(" Modelo cargado desde Registry.")
except:
    print(" Fallo de red/DFS. Activando modelo de emergencia con datos locales...")
    
    df_raw = spark.read.parquet("/opt/spark-data/processed/secop_ml_ready.parquet").limit(500)
    
    
    binarizer = Binarizer(threshold=100000000, inputCol="label", outputCol="label_bin")
    lr = LogisticRegression(labelCol="label_bin", featuresCol="features")
    
    pipeline_emergency = Pipeline(stages=[binarizer, lr])
    production_model = pipeline_emergency.fit(df_raw)
    print(" Modelo de emergencia listo para inferencia.")



RESPUESTA: Cargar desde el Registry permite centralizar la gobernanza, facilitar rollbacks y asegurar que producción use el modelo validado en experimentos.

## RETO 2 y 3 Inferencia

from pyspark.sql.functions import current_timestamp, lit


df_new = spark.read.parquet("/opt/spark-data/processed/secop_ml_ready.parquet")

# 2. Preparar el DataFrame para el Pipeline
df_scoring = df_new.select("features", "label") 

print("Generando predicciones batch...")

try:
 
    predictions_batch = production_model.transform(df_scoring)


    predictions_batch = predictions_batch.withColumn("prediction_timestamp", current_timestamp())


    print("Predicciones generadas con éxito:")
    predictions_batch.select("prediction", "probability", "prediction_timestamp").show(5, truncate=False)

except Exception as e:
    print(f" Error en la transformación: {e}")
    print("\nAnálisis: El Pipeline espera columnas específicas. Verificando columnas actuales:")
    print(df_scoring.columns)

## RETO 4 y 5: Monitoreo y Guardado

from pyspark.sql.functions import avg, count

# Estadísticas (Reto 4)
stats = predictions_batch.select(
    avg("prediction").alias("tasa_exito"),
    count("*").alias("total")
).collect()[0]

print("\n=== REPORTE DE PRODUCCIÓN ===")
print(f"Total procesado: {stats['total']:,}")
print(f"Tasa de detección: {stats['tasa_exito']*100:.2f}%")

# Guardar Resultados (Reto 5)
output_path = "/opt/spark-data/results/predicciones_notebook_12"
predictions_batch.write.mode("overwrite").parquet(output_path)

print(f"\n Archivo guardado en: {output_path}")

El sistema filtró el 7.71% que tienen el perfil de mayor riesgo o importancia por su valor.
